+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+
+- [add] Allow reading raw measurement data (#28)
+- [add] Temperature/Humidity: Expose `from_raw` constructor (#28)
+- [add] Impl `Copy` for all measurement data types (#28)
+
 ## 0.10.0 - 2020-10-23
 
 - [add] Provide non-blocking API for measurements (#16)